pyproject.toml
[tool.poetry]
name = "escro-ai-agent"
version = "0.1.0"
description = ""
authors = ["Karan Singh <pyrotank41@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.110.2"
streamlit = "^1.33.0"
beautifulsoup4 = "^4.12.3"
lxml = "^5.2.1"
html5lib = "^1.1"
jupyter = "^1.0.0"
requests = "^2.31.0"
llama-index = "^0.10.32"
trulens-eval = "^0.28.2"
python-dotenv = "^1.0.1"
watchdog = "^4.0.0"
llama-index-embeddings-huggingface = "^0.2.0"
llama-index-vector-stores-qdrant = "^0.2.8"
llama-index-llms-ollama = "^0.1.3"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
